+++ lib/acvp_vendor.c
@@ -173,18 +173,12 @@ out:
 static int acvp_vendor_validate_one(const struct acvp_testid_ctx *testid_ctx,
 				    const struct def_vendor *def_vendor)
 {
-	const struct acvp_ctx *ctx = testid_ctx->ctx;
-	const struct acvp_net_ctx *net;
 	struct json_object *resp = NULL, *data = NULL;
 	ACVP_BUFFER_INIT(buf);
 	int ret, ret2;
 	char url[ACVP_NET_URL_MAXLEN];
 
-	CKNULL_LOG(ctx, -EINVAL,
-		   "Vendor validation: authentication context missing\n");
-	net = &ctx->net;
-
-	CKINT(acvp_create_url(net, NIST_VAL_OP_VENDOR, url, sizeof(url)));
+	CKINT(acvp_create_url(NIST_VAL_OP_VENDOR, url, sizeof(url)));
 	CKINT(acvp_extend_string(url, sizeof(url), "/%u",
 				 def_vendor->acvp_vendor_id));
 
@@ -215,13 +209,11 @@ out:
 static int acvp_vendor_register(const struct acvp_testid_ctx *testid_ctx,
 				struct def_vendor *def_vendor)
 {
-	const struct acvp_ctx *ctx = testid_ctx->ctx;
-	const struct acvp_net_ctx *net = &ctx->net;
 	struct json_object *json_vendor = NULL;
 	int ret;
 	char url[ACVP_NET_URL_MAXLEN];
 
-	CKINT(acvp_create_url(net, NIST_VAL_OP_VENDOR, url, sizeof(url)));
+	CKINT(acvp_create_url(NIST_VAL_OP_VENDOR, url, sizeof(url)));
 
 	/* Build JSON object with the oe specification */
 	CKINT(acvp_vendor_build(def_vendor, &json_vendor));
@@ -243,7 +235,6 @@ static int acvp_vendor_validate_all(const struct acvp_testid_ctx *testid_ctx,
 {
 	const struct acvp_ctx *ctx = testid_ctx->ctx;
 	const struct acvp_opts_ctx *ctx_opts = &ctx->options;
-	const struct acvp_net_ctx *net = &ctx->net;
 	struct json_object *resp = NULL, *data = NULL, *array;
 	ACVP_BUFFER_INIT(buf);
 	unsigned int i;
@@ -251,7 +242,7 @@ static int acvp_vendor_validate_all(const struct acvp_testid_ctx *testid_ctx,
 	char url[ACVP_NET_URL_MAXLEN];
 	bool found = false;
 
-	CKINT(acvp_create_url(net, NIST_VAL_OP_VENDOR, url, sizeof(url)));
+	CKINT(acvp_create_url(NIST_VAL_OP_VENDOR, url, sizeof(url)));
 
 	ret2 = _acvp_process_retry_testid(testid_ctx, &buf, url);